+++ static/attraction.html
@@ -107,7 +107,7 @@
 </div>
 <div class="footer">COPYRIGHT © 2021 台北一日遊</div>
 
-<script src="/static/js/attraction.js"></script>
+<script  type="module" src="/static/js/attraction.js"></script>
   
 </body>
 </html>
