+++ static/index.html
@@ -101,7 +101,7 @@
       -->
   </div>
   <div class="footer">COPYRIGHT © 2021 台北一日遊</div>
-  <script src="/static/js/index.js"></script>
+  <script  type="module" src="/static/js/index.js"></script>
